Explicit casts to uint8_t in VT100 escapes.

+++ src/ansi.cpp
@@ -15,7 +15,7 @@
 //  Returning the escape array by value is not very efficient.
 //  Find out if there is a better implementation. 
 std::array<uint8_t, 10> term::ansi::go_to(const ScreenPosition& screen_pos) {
-    auto pos = utils::array_of<uint8_t>(27, 91, 0, 0, 0, 59, 0, 0, 0, 72); // x1b[row:colH
+    auto pos = utils::array_of<uint8_t>((uint8_t)27, (uint8_t)91, (uint8_t)0, (uint8_t)0, (uint8_t)0, (uint8_t)59, (uint8_t)0, (uint8_t)0, (uint8_t)0, (uint8_t)72); // x1b[row:colH
     const auto row_start = pos.data() + 3, col_start = pos.data() + 6;
  
     std::to_chars((char*)row_start, (char*)row_start + 2, screen_pos.row);
